Update index powershell script
We need this to create the data for the DB to look through. Nomads is a
featured mod on faf so we can get its units from the same folder, once a
nomads game has been played. The file structure is the same so we just
need to add it as a directory.

Also this hasnt been tested, but it should just work. Right? right??