Add support for the Nomads faction (#89)
* Add Nomads unit icons

Pretty much what it says on the tin, we need these to show the units.

* Add Nomads ID to DB

Pretty much what it says on the tin, the faction needs an ID

* Add Nomads to faction IDs lists

Various places have this hard coded so we need to update it to also have
nomads.

* Update cleaner to have nomads

Pretty much what it says on the tin, mostly its adding the long list of
unit IDs, and then a few faction specific variables that are hard coded.

* Add placeholder Nomads image

right now its the uef copied over but we can add a nice nomads version
later.

* Update css to include Nomads

Currently its actually linking to a uef image but we can replace that
later easily without touching this file again.

* Update index powershell script

We need this to create the data for the DB to look through. Nomads is a
featured mod on faf so we can get its units from the same folder, once a
nomads game has been played. The file structure is the same so we just
need to add it as a directory.

Also this hasnt been tested, but it should just work. Right? right??

* Remove Unused Icons

a few icons werent belonging to any units used in the game, so they have
been removed.

* Add artillery satellite to list

Previously it was missing and its icon was wrong, now its in the
whitelist, since its an actual unit (though a bit weird since its an
assisting artillery unit, but you probably want to view its stats
anyway)

* Resize Nomads icons

After adding them, some were the wrong size so now they are all 64x64

* Update Nomads Image

Pretty much this. using pro photoshop skills its as if the guy made it
himself!

* Fix Typo In cleaner.py

Presumably it could still work like this since its a sort order but good
to have it actually correct.

* Remove Campaign unit from DB

XNC0001 is a special version of the orbital frigate used as a civilian
structure, so it doesnt need to be added into the DB.

* Add strategic icons for Nomads

Hopefully these are named correctly and work - these are needed for the
compact view which shows units by strategic instead of build icon, and
the icons need to be coloured so we cant just reuse them.

* Add missing strategic icons

Some were missing due to the other icons being copied from cybran so
Nomads things like the heavy t3 tank werent there. now they are. great!

* Add glow to unit icons

This is poorly tested so hopefully it should just work, but all this
does is add a slight glow around the units to make them look more
consistent.

* Add empty icon

This is pretty cursed actually but apparently you can just put this in
and it will work as the default icon because theres no default handling
code but it sort of breaks weird. oh well. bugs on top of bugs it is!

* Update css with Spooky

Spooky did this code. This updates the css file to make nomads able to
fit in the layout better.

* Fix typos, small errors and tweaks with Spooky

Spooky did this code.

Also i am ashamed to say i put syntax errors into the py script since i
didnt test it.

* Update DB to 3714 + Nomads with spooky

Spooky did this code. This is just whats generated out of the
powershells script.

* Update categories to support nomads units

Hopefully this should work, this is what categorises the units in the
other view of the database.

A bit cursed but you know.

* Add Nomads T2 shield generator category

actually this is cursed since this is fixing a typo but we do it like
this so we dont have to wait for release of the nomads mod to update the
DB. yeah.

* Fix typo

Commas are missing! also added one at the end so the next time i
duplicate a line i dont get hit by the curse.

* Fix categories missing

There was a typo in the railgun carrier blueprint id, and the tank
destroyer was missing in its list, so should be fixed now.

* Add glow to factory icons

The HQs were forgotten in the glow adding operation, since they were
listed at the bottom!