Misc bug fixing (#91)
* Add Nomads ASF to database

This was previously missing for some unknown reason, but is fixed now!

* Add Subcommander icons

for some reason the blueprints had those icons changed to subcommander
at some point, making the db missing those new icons. they are added
now.

* Fix Nomads destroyer icon missing

Pretty much what it says on the tin, this was forgotten about before!